Split cdc up into multiple files
